* Exportada funcao AchaIndiceCombo de GtkUtils para uso pelo programa * Agora desmarca o material em uso ao sair de producao, evitando que o operador deixe marcada uma bobina que tenha terminado * Alterado script de compilacao para exibir versao e data ao termino da compilacao * Agora eh possivel ver / registrar produtos na tela de materiais + Adicionada opcao para movimentar materiais entre setores (Pintura, Solda, Local) * Corrigido problema de precisao (casas decimais) dos dados enviados nas mensagens + Adicionado controle de materiais para a perfiladeira de porta-palete normal * Alterado combo de tipos de materiais para listar apenas os tipos utilizados nas maquinas (1 - materia prima, 3 - produto em processo, 5 - subproduto (sucata) ) + Adicionado filtro para exibir materiais por tipo na tela de materiais + Adicionada mensagem de transferencia de materiais entre depositos

+++ main.c
@@ -704,11 +704,11 @@ void Board_Init(BoardStatus *bs)
 }
 
 // Funcao que converte float para string, trocando a virgula do separador por ponto
-char *floatToString(char *dst, float val)
+char *floatToString(char *dst, double val)
 {
     char *pos_comma;
 
-    sprintf(dst, "%f", val);
+    sprintf(dst, "%.9lf", val);
     pos_comma = strchr(dst, ',');
     if(pos_comma != NULL) *pos_comma = '.';
 
@@ -716,9 +716,9 @@ char *floatToString(char *dst, float val)
 }
 
 // Funcao que converte float para string, trocando a virgula do separador por ponto
-float StringToFloat(char *val)
+double StringToFloat(char *val)
 {
-	float result;
+	double result;
     char *pos_comma;
 
     // Se string nula, retorna zero
